cmd/record: cancel ctx for spawned sub-shell when pty closes
When the user presses exit or ctrl-d, the pty will be closed and
io.Copy(os.Stdout, ptmx) will unblock and we will call cancelCtx

This should get rid of the scenario where users press ctrl-d or exit and
the terminal just hangs blocked waiting for c.Wait() to unblock.

We force c.Wait() to unblock by cancelling the context.