history: selected commands are sorted chronologically by time.
i:e the oldest commands ( smallest unix timestamp ) will show up first.

This is the reverse order from where how the history is presented.

Usually, devs want to create a runbook from most recent commands, so we
show history in that order.

But once selected, we want to respect the chronology of time.