shell/zsh: use COMMA as a delimter when passing in SAVVY_RUNBOOK_COMMANDS
When using the literal `,` as a delimeter, we can't split on `,` to
reconstruct the sequence of commands as an array in zsh.

Commands like grep --exlcude-dir{.git,foo,etc} include false positives.

One alternative is to escape those commands, but using COMMA is simpler.

We also take this opportunity to use `@s` instead of just `s` while
splitting the string.

@s preserves the spacing and quotes around string